Update Checkstyle thresholds after a successful Travis build
* Use GitHub API to push changes back to origin

We can't push directly from Travis because Travis makes a shallow clone
of the repo.  We could convert the shallow clone to a full clone using
"git fetch --unshallow" when we detect the need to push changes back to
the origin, but this can take an extremely long time due to the size of
the TripleA repo (it takes almost 10 minutes to make a full clone on my
machine).

Instead, we use the GitHub API to update the changed gradle.properties
file.  The only need for Git using this method is to calculate the blob
SHA of the original file.  (The Git SHA is not simply the SHA-1 of the
content; there is a specific header Git prepends to the content before
calculating the hash.)

* Use TRAVIS_BUILD_NUMBER in commit message

TAGGED_VERSION is only available once the deploy phase has started.  The
Checkstyle threshold update script runs before the deploy phase, so we
can't use this variable.  Instead we use TRAVIS_BUILD_NUMBER, which
still uniquely identifies which build triggered the commit.

We could have constructed TAGGED_VERSION internal to the script, but it
seems like a bad smell to duplicate this logic (especially since it may
be changed in the near future).

* Skip Checkstyle threshold update if build is not permitted

A Checkstyle threshold update is only performed when the build
satisifies the following critieria:

- The build is not a pull request.
- The build was triggered on branch 'master'.
- The build was triggered on repo 'triplea-game/triplea'.