Checkstyle: Rename fields and methods to match code style
I introduced new Checkstyle violations in the previous commit, which
causes a build failure due to the newly-added enforcement of Checkstyle
violation thresholds.  Unfortunately, those violations were due to the
legacy naming convention used in the MapProperties class to facilitate
reflection.

This change not only removes the violations added in the previous
commit, but removes all naming convention violations in the
MapProperties class.  This required a small modification in
MapPropertiesWrapper to distinguish between a property name and the name
of the field that backs the property in order to use the recommended
naming convention.